requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.1
django-filter==2.1.0
django-stubs==0.12.1
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
idna==2.8
mypy==0.711
mypy-extensions==0.4.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
gunicorn==19.7.1